+++ Tests/SentryTests/SentryOptionsTest.m
@@ -518,6 +518,7 @@
 - (void)testNSNull_SetsDefaultValue
 {
     SentryOptions *options = [[SentryOptions alloc] initWithDict:@{
+        @"urlSession" : [NSNull null],
         @"dsn" : [NSNull null],
         @"enabled" : [NSNull null],
         @"debug" : [NSNull null],
@@ -617,6 +618,7 @@
     XCTAssertEqualObjects([self getDefaultInAppIncludes], options.inAppIncludes);
     XCTAssertEqual(@[], options.inAppExcludes);
     XCTAssertNil(options.urlSessionDelegate);
+    XCTAssertNil(options.urlSession);
     XCTAssertEqual(YES, options.enableSwizzling);
     XCTAssertEqual([NSSet new], options.swizzleClassNameExcludes);
     XCTAssertEqual(YES, options.enableFileIOTracing);
@@ -1259,6 +1261,16 @@
                                                       didFailWithError:&error];
     XCTAssertNil(error);
     return sentryOptions;
+}
+
+- (void)testURLSession
+{
+    NSURLSession *urlSession = [NSURLSession
+        sessionWithConfiguration:[NSURLSessionConfiguration ephemeralSessionConfiguration]];
+
+    SentryOptions *options = [self getValidOptions:@{ @"urlSession" : urlSession }];
+
+    XCTAssertNotNil(options.urlSession);
 }
 
 - (void)testUrlSessionDelegate