+++ Sources/Swift/Extensions/URLSessionTaskExtensions.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+public extension URLSessionTask {
+
+    @objc
+    func getGraphQLOperationName() -> String? {
+        guard originalRequest?.value(forHTTPHeaderField: "Content-Type") == "application/json" else { return nil }
+        guard let requestBody = originalRequest?.httpBody else { return nil }
+
+        let requestInfo = try? JSONDecoder().decode(GraphQLRequest.self, from: requestBody)
+
+        return requestInfo?.operationName
+    }
+
+}
+
+private struct GraphQLRequest: Decodable {
+    let operationName: String
+}