+++ Sources/Sentry/SentrySessionReplayIntegration.m
@@ -28,17 +28,6 @@ SentrySessionReplayIntegration ()
 - (void)newSceneActivate;
 @end
 
-API_AVAILABLE(ios(16.0), tvos(16.0))
-@interface
-SentryViewPhotographer (SentryViewScreenshotProvider) <SentryViewScreenshotProvider>
-@end
-
-API_AVAILABLE(ios(16.0), tvos(16.0))
-@interface
-SentryOnDemandReplay (SentryReplayMaker) <SentryReplayMaker>
-
-@end
-
 @implementation SentrySessionReplayIntegration {
     BOOL _startedAsFullSession;
     SentryReplayOptions *_replayOptions;
@@ -150,6 +139,7 @@ SentryOnDemandReplay (SentryReplayMaker) <SentryReplayMaker>
 
 - (void)uninstall
 {
+    [self stop];
 }
 
 - (BOOL)shouldReplayFullSession:(CGFloat)rate