+++ Sources/SentryCrash/Recording/SentryCrashCachedData.c
@@ -24,6 +24,7 @@
 //
 
 #include "SentryCrashCachedData.h"
+#include "SentryInternalCDefines.h"
 
 // #define SentryCrashLogger_LocalLevel TRACE
 #include "SentryCrashLogger.h"
@@ -54,7 +55,7 @@ static _Atomic(int) g_semaphoreCount;
 static bool g_hasThreadStarted = false;
 
 static void
-updateThreadList(void)
+updateThreadList(void) SENTRY_DISABLE_THREAD_SANITIZER("Known data race to fix")
 {
     const task_t thisTask = mach_task_self();
     int oldThreadsCount = g_allThreadsCount;
@@ -129,6 +130,7 @@ updateThreadList(void)
 
 static void *
 monitorCachedData(__unused void *const userData)
+    SENTRY_DISABLE_THREAD_SANITIZER("Known data race to fix")
 {
     static int quickPollCount = 4;
     usleep(1);
@@ -149,6 +151,7 @@ monitorCachedData(__unused void *const userData)
 
 void
 sentrycrashccd_init(int pollingIntervalInSeconds)
+    SENTRY_DISABLE_THREAD_SANITIZER("Known data race to fix")
 {
     if (g_hasThreadStarted == true) {
         return;
@@ -211,6 +214,7 @@ sentrycrashccd_getAllThreads(int *threadCount)
 
 const char *
 sentrycrashccd_getThreadName(SentryCrashThread thread)
+    SENTRY_DISABLE_THREAD_SANITIZER("Known data race to fix")
 {
     if (g_allThreadNames != NULL) {
         for (int i = 0; i < g_allThreadsCount; i++) {