+++ CHANGELOG.md
@@ -6,14 +6,13 @@
 
 - Ignore SentryFramesTracker thread sanitizer data races (#3922)
 - Handle no releaseName in WatchDogTerminationLogic (#3919)
+- Add SIGTERM support (#3895)
 
 ### Improvements
 
 - Remove not needed lock for logging (#3934)
 
 ## 8.25.0
-
-- Add SIGTERM support ([#3895](https://github.com/getsentry/sentry-cocoa/pull/3895))
 
 ### Features