+++ CHANGELOG.md
@@ -7,10 +7,15 @@
 - Ignore SentryFramesTracker thread sanitizer data races (#3922)
 - Handle no releaseName in WatchDogTerminationLogic (#3919)
 - Add SIGTERM support (#3895)
+- Stop SessionReplay when closing SDK (#3941)
 
 ### Improvements
 
 - Remove not needed lock for logging (#3934)
+- Session replay Improvements (#3877)
+  - Use image average color and text font color to redact session replay 
+  - Removed iOS 16 restriction from session replay
+  - Performance improvement 
 
 ## 8.25.0